backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
pymongo==4.4.0
bcrypt==4.0.1
PyJWT==2.8.0
numpy==1.25.0
scikit-learn==1.3.2
sentence-transformers==2.2.2
feedparser==6.0.10
python-dotenv==1.0.0
